backend/requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-RESTX==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Security
python-dotenv==1.0.0
bcrypt==4.1.2

# Utilities
marshmallow==3.20.1
python-dateutil==2.8.2
python-slugify==8.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
factory-boy==3.3.0

# Development
black==23.12.1
flake8==7.0.0

# Production server
gunicorn==21.2.0